fix(frontend): fix bug with xstate init/re-init
