+++ frontend/app/routes/protected/person-case/contact-information.tsx
@@ -26,8 +26,8 @@ import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/protected/person-case/layout';
-import type { ContactInformationData } from '~/routes/protected/person-case/state-machine';
 import { getStateRoute, loadMachineActor } from '~/routes/protected/person-case/state-machine';
+import type { ContactInformationData } from '~/routes/protected/person-case/types';
 
 const log = LogFactory.getLogger(import.meta.url);
