+++ frontend/app/routes/protected/person-case/personal-info.tsx
@@ -23,8 +23,8 @@ import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/protected/person-case/layout';
-import type { PersonalInfoData } from '~/routes/protected/person-case/state-machine';
 import { getStateRoute, loadMachineActor } from '~/routes/protected/person-case/state-machine';
+import type { PersonalInfoData } from '~/routes/protected/person-case/types';
 import { REGEX_PATTERNS } from '~/utils/regex-utils';
 
 const log = LogFactory.getLogger(import.meta.url);
