+++ frontend/app/routes/protected/person-case/previous-sin.tsx
@@ -26,8 +26,8 @@ import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/protected/person-case/layout';
-import type { PreviousSinData } from '~/routes/protected/person-case/state-machine';
 import { getStateRoute, loadMachineActor } from '~/routes/protected/person-case/state-machine';
+import type { PreviousSinData } from '~/routes/protected/person-case/types';
 import { formatSin, isValidSin, sinInputPatternFormat } from '~/utils/sin-utils';
 
 const log = LogFactory.getLogger(import.meta.url);
