+++ frontend/app/routes/protected/person-case/privacy-statement.tsx
@@ -19,8 +19,8 @@ import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/protected/person-case/layout';
-import type { PrivacyStatementData } from '~/routes/protected/person-case/state-machine';
 import { createMachineActor, getStateRoute, loadMachineActor } from '~/routes/protected/person-case/state-machine';
+import type { PrivacyStatementData } from '~/routes/protected/person-case/types';
 
 const log = LogFactory.getLogger(import.meta.url);
