+++ frontend/app/routes/protected/person-case/request-details.tsx
@@ -28,8 +28,8 @@ import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/protected/person-case/layout';
-import type { RequestDetailsData } from '~/routes/protected/person-case/state-machine';
 import { getStateRoute, loadMachineActor } from '~/routes/protected/person-case/state-machine';
+import type { RequestDetailsData } from '~/routes/protected/person-case/types';
 
 const log = LogFactory.getLogger(import.meta.url);
